Create HeatmapExample.html
This is an example of an implementation using Google Maps API. The HTML/javascript program pulls GeoJson coordinates from a GeoJson file containing coordinates of earthquake data. Our program parses the GeoJson file, populates an array with all of the coordinates, and generates a heatmap based on that array. Now that we have a working heatmap example that parses GeoJson correctly, we will edit this to read our own GeoJson and produce a heatmap from our results.